claude: extract version info into module
