+++ 3p/claude/cook.mk
@@ -1,0 +1,6 @@
+claude-latest: private .PLEDGE = stdio rpath wpath cpath inet dns exec proc
+claude-latest: private .INTERNET = 1
+claude-latest: $(lua_bin)
+	$(lib_lua) 3p/claude/latest.lua
+
+.PHONY: claude-latest
