+++ 3p/claude/version.lua
@@ -1,0 +1,9 @@
+return {
+  version = "2.0.74",
+  url = "https://storage.googleapis.com/claude-code-dist-86c565f3-f756-42ad-8dfa-d59b1c096819/claude-code-releases/{version}/{platform}/claude",
+  platforms = {
+    ["linux-x64"] = {
+      sha = "43065ff86a1b952225e42042bf4dfe9f6b72ff8ed91686a23add5396b1a11e80",
+    },
+  },
+}
